+++ src/main/java/fr/utbm/fisa/communicationapi/infrastructure/repositories/GroupRepository.java
@@ -3,6 +3,6 @@ package fr.utbm.fisa.communicationapi.infrastructure.repositories;
 import org.springframework.data.repository.CrudRepository;
 import fr.utbm.fisa.communicationapi.infrastructure.entities.Group;
 
-public interface GroupRepository extends CrudRepository<Group, Integer> {
+public interface GroupRepository extends CrudRepository<Group, Long> {
 
 }
